+++ src/components/button/button.tsx
@@ -1,17 +1,15 @@
-import * as React from "react";
 import { twMerge } from "tailwind-merge";
 import { type VariantProps } from "class-variance-authority";
 import { buttonVariants } from "./variants";
+import { ComponentProps, FC } from "react";
 
-type ButtonProps = React.ComponentProps<"button"> &
+type ButtonProps = ComponentProps<"button"> &
   VariantProps<typeof buttonVariants> & {
     fullWidth?: boolean;
     isCircular?: boolean;
-    /** Needed for icon-only buttons */
-    "aria-label"?: string;
   };
 
-export const Button: React.FC<ButtonProps> = ({
+export const Button: FC<ButtonProps> = ({
   variant = "primary",
   size = "default",
   fullWidth = false,