+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [2.13.0](https://github.com/ost-cas-fea-25-26/pp-design-system/compare/v2.12.1...v2.13.0) (2025-11-09)
+
+
+### Features
+
+* add app-shell composition with header and logo-link ([#103](https://github.com/ost-cas-fea-25-26/pp-design-system/issues/103)) ([736ce36](https://github.com/ost-cas-fea-25-26/pp-design-system/commit/736ce36dad06d00870b51a098a0f890632d2b313))
+
+## [2.12.1](https://github.com/ost-cas-fea-25-26/pp-design-system/compare/v2.12.0...v2.12.1) (2025-11-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency lucide-react to ^0.553.0 ([#115](https://github.com/ost-cas-fea-25-26/pp-design-system/issues/115)) ([d3313b9](https://github.com/ost-cas-fea-25-26/pp-design-system/commit/d3313b9ba547816660f66555f50694ed1286cfb3))
+
 # [2.12.0](https://github.com/ost-cas-fea-25-26/pp-design-system/compare/v2.11.1...v2.12.0) (2025-11-05)